SBGrid Fix
SBGrid fix for color error

After the Python 3 upgrade, this library was complaining about a color being false and bombing out. Our template we were passing didn't change. The section of code bombing out was already had a fallback for color being None, and it was expanded to also handle color being false.